% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_harmonise.R
\name{eirer_crs}
\alias{eirer_crs}
\title{The coordinate reference system used throughout EireR}
\usage{
eirer_crs()
}
\value{
Integer `2157L`
}
\description{
All `get_*()` functions return data in ETRS89 / Irish Transverse Mercator
EPSG:2157. Distances and areas in this CRS are in metres.
}
